Observe2Agent/qa-automation-platform/backend/services/video_analyzer.py
"""
Video analysis service using Google Gemini to understand business process workflows.
Extracts insights, detects systems, and identifies process steps from videos.
"""

import logging
from typing import List, Dict, Any, Optional
import json
from datetime import datetime
from pathlib import Path

try:
    import google.generativeai as genai
except ImportError:
    genai = None

from config import settings

logger = logging.getLogger(__name__)


class VideoAnalyzer:
    """
    Analyzes business process videos using Gemini multimodal capabilities.
    Identifies workflow steps, systems involved, and generates detailed insights.
    """

    def __init__(self):
        """Initialize the video analyzer with Gemini configuration."""
        if genai and settings.GOOGLE_API_KEY:
            genai.configure(api_key=settings.GOOGLE_API_KEY)
        self.model_name = settings.GEMINI_MODEL
        self.temperature = settings.GEMINI_TEMPERATURE
        self.max_tokens = settings.GEMINI_MAX_TOKENS

    def analyze_video(self, video_path: str) -> Dict[str, Any]:
        """
        Analyze a video file to extract business process information.

        Args:
            video_path: Path to the video file

        Returns:
            Dictionary containing analysis results with steps, systems, and insights
        """
        logger.info(f"Starting video analysis for: {video_path}")

        try:
            # Validate video file exists
            if not Path(video_path).exists():
                logger.error(f"Video file not found: {video_path}")
                return {
                    "success": False,
                    "error": f"Video file not found: {video_path}",
                }

            # For now, return mock data as Gemini video upload API requires special setup
            # In production, this would use genai.upload_file() and process_video()
            analysis_result = self._get_mock_analysis(video_path)

            logger.info("Video analysis completed successfully")
            return analysis_result

        except Exception as e:
            logger.error(f"Error analyzing video: {str(e)}")
            return {
                "success": False,
                "error": str(e),
            }

    def _get_mock_analysis(self, video_path: str) -> Dict[str, Any]:
        """
        Generate mock video analysis for development/testing.
        In production, this would be replaced with actual Gemini analysis.

        Args:
            video_path: Path to the video file

        Returns:
            Dictionary with simulated analysis results
        """
        return {
            "success": True,
            "video_path": video_path,
            "video_duration": 120.5,
            "frames_extracted": 25,
            "workflow_steps": [
                {
                    "step_number": 1,
                    "title": "Open SAP Application",
                    "description": "Launch SAP ERP system and wait for login screen",
                    "timestamp": 0.0,
                    "duration": 8.5,
                    "action_type": "click",
                    "system": "SAP",
                    "ui_elements": ["SAP Logo", "Login Form"],
                },
                {
                    "step_number": 2,
                    "title": "Enter Credentials",
                    "description": "Input username and password in login form",
                    "timestamp": 8.5,
                    "duration": 12.0,
                    "action_type": "input",
                    "system": "SAP",
                    "ui_elements": ["Username Field", "Password Field", "Login Button"],
                },
                {
                    "step_number": 3,
                    "title": "Navigate to Purchase Order Module",
                    "description": "Access the procurement module from main menu",
                    "timestamp": 20.5,
                    "duration": 10.2,
                    "action_type": "navigate",
                    "system": "SAP",
                    "ui_elements": ["Main Menu", "Procurement Module", "Purchase Order Option"],
                },
                {
                    "step_number": 4,
                    "title": "Create New Purchase Order",
                    "description": "Click create button and fill in basic details",
                    "timestamp": 30.7,
                    "duration": 25.3,
                    "action_type": "click",
                    "system": "SAP",
                    "ui_elements": ["Create Button", "Form Fields", "Save Button"],
                },
                {
                    "step_number": 5,
                    "title": "Submit for Approval",
                    "description": "Send PO to approver and confirm submission",
                    "timestamp": 56.0,
                    "duration": 15.8,
                    "action_type": "click",
                    "system": "SAP",
                    "ui_elements": ["Submit Button", "Confirmation Dialog"],
                },
            ],
            "systems_detected": [
                {
                    "name": "SAP ERP",
                    "system_type": "ERP",
                    "confidence": 0.95,
                    "first_detected_at": 0.0,
                    "ui_elements": [
                        "SAP Logo",
                        "Fiori Interface",
                        "Transaction Code Field",
                    ],
                    "version": "S/4HANA 2021",
                },
                {
                    "name": "Email Client",
                    "system_type": "Email",
                    "confidence": 0.78,
                    "first_detected_at": 85.0,
                    "ui_elements": ["Outlook Logo", "Email Compose"],
                },
            ],
            "data_extraction_patterns": [
                {
                    "field_name": "Purchase_Order_Number",
                    "pattern": "PO-[0-9]{8}",
                    "confidence": 0.92,
                    "example": "PO-20240112",
                },
                {
                    "field_name": "Vendor_Name",
                    "pattern": "Text input field",
                    "confidence": 0.88,
                    "example": "Acme Supplies Inc.",
                },
                {
                    "field_name": "Amount",
                    "pattern": "Currency format",
                    "confidence": 0.91,
                    "example": "$5,000.00",
                },
            ],
            "process_summary": "SAP Purchase Order creation process involving login, navigation to procurement module, creating a new PO with vendor and amount details, and submitting for approval via email.",
            "estimated_execution_time": 56.0,
            "success_indicators": [
                "PO number generated",
                "Status shows as pending approval",
                "Email confirmation sent",
            ],
            "analysis_timestamp": datetime.utcnow().isoformat(),
        }

    def extract_frames(self, video_path: str, interval: int = 5) -> List[str]:
        """
        Extract frames from video at specified intervals.

        Args:
            video_path: Path to the video file
            interval: Extract frame every N seconds

        Returns:
            List of paths to extracted frame images
        """
        logger.info(f"Extracting frames from {video_path} every {interval} seconds")

        # In production, this would use OpenCV or FFmpeg to extract frames
        # For now, return mock frame paths
        frame_paths = [
            f"/tmp/frames/frame_00_{i*interval:03d}.png" for i in range(10)
        ]

        logger.info(f"Extracted {len(frame_paths)} frames")
        return frame_paths

    def detect_ui_elements(self, frame_path: str) -> List[Dict[str, Any]]:
        """
        Detect UI elements in a frame using Gemini vision.

        Args:
            frame_path: Path to the frame image

        Returns:
            List of detected UI elements with coordinates and labels
        """
        logger.info(f"Detecting UI elements in: {frame_path}")

        # Mock implementation
        return [
            {
                "element_type": "button",
                "label": "Login",
                "coordinates": {"x": 100, "y": 200, "width": 80, "height": 40},
                "confidence": 0.94,
            },
            {
                "element_type": "input",
                "label": "Username",
                "coordinates": {"x": 50, "y": 100, "width": 200, "height": 30},
                "confidence": 0.91,
            },
        ]

    def identify_process_flow(self, steps: List[Dict[str, Any]]) -> Dict[str, Any]:
        """
        Identify the overall process flow from detected steps.

        Args:
            steps: List of workflow steps

        Returns:
            Process flow analysis with dependencies and patterns
        """
        logger.info(f"Analyzing process flow from {len(steps)} steps")

        return {
            "process_type": "sequential",
            "has_loops": False,
            "has_conditionals": False,
            "estimated_duration": sum(step.get("duration", 0) for step in steps),
            "critical_steps": [0, 3],  # Login and Submit steps are critical
            "dependencies": {
                "1": ["0"],  # Credentials depend on login
                "2": ["1"],  # Navigation depends on credentials
                "3": ["2"],  # Create PO depends on navigation
                "4": ["3"],  # Submit depends on creation
            },
            "data_flow": [
                {
                    "source": "login",
                    "target": "po_creation",
                    "data": "user_session",
                }
            ],
        }
