Observe2Agent/qa-automation-platform/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.0
google-cloud-documentai==1.17.1
google-cloud-storage==2.10.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
aiofiles==23.2.1
requests==2.31.0
sqlalchemy==2.0.23
httpx==0.25.1
python-json-logger==2.0.7
